+++ R/app-driver.R
@@ -127,11 +127,13 @@ NULL
 #'
 #' @param ... Must be empty. Allows for parameter expansion.
 #' @param timeout Amount of time to wait before giving up (milliseconds).
+#'   Defaults to the resolved value of `AppDriver$new(timeout=)`
 #' @param cran Should these expectations be verified on CRAN? By default,
 #'        they are not because snapshot tests tend to be fragile
 #'        because they often rely on minor details of dependencies.
 #' @param wait_ Wait until all reactive updates have completed?
 #' @param timeout_ Amount of time to wait before giving up (milliseconds).
+#'   Defaults to the resolved value of `AppDriver$new(timeout=)`
 #' @param hash_images If `TRUE`, images will be hashed before being returned.
 #'   Otherwise, all images will return their full data64 encoded value.
 #' @param screenshot_args This named list of arguments is passed along to
@@ -166,7 +168,8 @@ AppDriver <- R6Class( # nolint
     counter = "<Count>",
     shiny_url = "<Url>",
 
-    timeout_stepsize = 1000, # milliseconds
+    load_timeout = 15 * 1000, # Time for shiny app to load; ms
+    timeout = 1000, # Default timeout value; ms
 
     logs = list(), # List of all log messages added via `$log_message()`
 
@@ -215,20 +218,19 @@ AppDriver <- R6Class( # nolint
     #'   For apps that use R's random number generator, this can make their
     #'   behavior repeatable.
     #' @param load_timeout How long to wait for the app to load, in ms.
-    #'   This includes the time to start R. Defaults to 15s when running
-    #'   locally and 30s when running on CI.
-    #' @param timeout_stepsize Number of milliseconds that a single timeout step represents.
+    #'   This includes the time to start R. Defaults to 15s.
     #'
-    #'   If `timeout_stepsize` is missing (default),
-    #'   `getOption("shinytest2.timeout.stepsize")` will be used. If no option
-    #'   value value is found, a default value of `2 * 1000` will be used on
-    #'   CI otherwise `timeout_stepsize` will be set to `1000` milliseconds.
+    #'   If `load_timeout` is missing, the first numeric value found will be used:
+    #'   * R option `options(shinytest2.load_timeout=)`
+    #'   * System environment variable `SHINYTEST2_LOAD_TIMEOUT`
+    #'   * `15 * 1000` (15 seconds)
+    #' @param timeout Default number of milliseconds for any `timeout` or
+    #'   `timeout_` parameter in the `AppDriver` class. Defaults to 4s.
     #'
-    #'   Having the different mechanisms to initialize the timeout_stepsize
-    #'   allows for users to:
-    #'   * set a default timeout directly (`AppDriver$new(timeout_stepsize = 500)`)
-    #'   * set a default timeout indirectly before creating a new `AppDriver` (`options(shinytest2.timeout.stepsize = 500)`)
-    #'   * or naturally allow for more execution time on CI (when `as.logical(Sys.getenv("CI"))` is `TRUE`).
+    #'   If `timeout` is missing, the first numeric value found will be used:
+    #'   * R option `options(shinytest2.timeout=)`
+    #'   * System environment variable `SHINYTEST2_TIMEOUT`
+    #'   * `4 * 1000` (4 seconds)
     #' @param wait If `TRUE`, `$wait_for_idle(duration = 200, timeout = load_timeout)`
     #'   will be called once the app has connected a new session, blocking until the
     #'   Shiny app is idle for 200ms.
@@ -276,8 +278,8 @@ AppDriver <- R6Class( # nolint
       name = NULL,
       variant = missing_arg(),
       seed = NULL,
-      load_timeout = self$get_timeout(15),
-      timeout_stepsize = missing_arg(),
+      load_timeout = missing_arg(),
+      timeout = missing_arg(),
       wait = TRUE,
       screenshot_args = missing_arg(),
       expect_values_screenshot_args = TRUE,
@@ -295,7 +297,7 @@ AppDriver <- R6Class( # nolint
         app_dir = app_dir,
         ...,
         load_timeout = load_timeout,
-        timeout_stepsize = timeout_stepsize,
+        timeout = timeout,
         wait = wait,
         expect_values_screenshot_args = expect_values_screenshot_args,
         screenshot_args = screenshot_args,
@@ -392,7 +394,7 @@ AppDriver <- R6Class( # nolint
     set_inputs = function(
       ...,
       wait_ = TRUE,
-      timeout_ = self$get_timeout(4),
+      timeout_ = missing_arg(),
       allow_no_input_binding_ = FALSE,
       priority_ = c("input", "event")
     ) {
@@ -420,7 +422,7 @@ AppDriver <- R6Class( # nolint
     #' # Upload file to input named: file1
     #' app$upload_file(file1 = tmpfile)
     #' }
-    upload_file = function(..., wait_ = TRUE, timeout_ = self$get_timeout(4)) {
+    upload_file = function(..., wait_ = TRUE, timeout_ = missing_arg()) {
       app_upload_file(self, private, ..., wait_ = wait_, timeout_ = timeout_)
     },
 
@@ -846,7 +848,7 @@ AppDriver <- R6Class( # nolint
       script = missing_arg(),
       ...,
       file = missing_arg(),
-      timeout = self$get_timeout(4),
+      timeout = missing_arg(),
       pre_snapshot = NULL,
       cran = FALSE
     ) {
@@ -922,7 +924,7 @@ AppDriver <- R6Class( # nolint
       script = missing_arg(),
       ...,
       file = missing_arg(),
-      timeout = self$get_timeout(4)
+      timeout = missing_arg()
     ) {
       app_get_js(
         self, private,
@@ -966,7 +968,7 @@ AppDriver <- R6Class( # nolint
       script = missing_arg(),
       ...,
       file = missing_arg(),
-      timeout = self$get_timeout(4)
+      timeout = missing_arg()
     ) {
       app_run_js(
         self, private,
@@ -1175,7 +1177,8 @@ AppDriver <- R6Class( # nolint
     #' resizing the window.)
     #' @param duration How long Shiny must be idle (in ms) before unblocking the
     #'   R session.
-    #' @param timeout How often to check for the condition, in ms.
+    #' @param timeout_ Amount of time to wait before giving up (milliseconds).
+    #'   Defaults to the resolved value of `AppDriver$new(timeout=)`
     #' @return `invisible(self)` if Shiny stabilizes within the `timeout`.
     #'   Otherwise an error will be thrown
     #' @examples
@@ -1208,7 +1211,7 @@ AppDriver <- R6Class( # nolint
     #' # Will not be equal
     #' identical(pre_value, post_value)
     #' }
-    wait_for_idle = function(duration = 500, timeout = self$get_timeout(4)) {
+    wait_for_idle = function(duration = 500, timeout = missing_arg()) {
       app_wait_for_idle(self, private, duration = duration, timeout = timeout)
     },
 
@@ -1225,7 +1228,8 @@ AppDriver <- R6Class( # nolint
     #'   Only one of these parameters may be used.
     #' @param ignore List of possible values to ignore when checking for
     #'   updates.
-    #' @param timeout How long to wait (in ms) before throwing an error.
+    #' @param timeout_ Amount of time to wait before giving up (milliseconds).
+    #'   Defaults to the resolved value of `AppDriver$new(timeout=)`
     #' @param interval How often to check for the condition, in ms.
     #' @return Newly found value
     #' @examples
@@ -1276,7 +1280,7 @@ AppDriver <- R6Class( # nolint
       output = missing_arg(),
       export = missing_arg(),
       ignore = list(NULL, ""),
-      timeout = self$get_timeout(4),
+      timeout = missing_arg(),
       interval = 400
     ) {
       app_wait_for_value(
@@ -1297,6 +1301,7 @@ AppDriver <- R6Class( # nolint
     #'   value](https://developer.mozilla.org/en-US/docs/Glossary/Truthy) or a
     #'   timeout error will be thrown.
     #' @param timeout How long the script has to return a `true`thy value, in ms.
+    #'   Defaults to the resolved value of `AppDriver$new(timeout=)`
     #' @param interval How often to check for the condition, in ms.
     #' @return `invisible(self)` if expression evaluates to `true` without error
     #'   within the timeout. Otherwise an error will be thrown
@@ -1316,8 +1321,17 @@ AppDriver <- R6Class( # nolint
     #' app$run_js(file = "complicated_file.js")
     #' app$wait_for_js("complicated_condition();")
     #' }
-    wait_for_js = function(script, timeout = self$get_timeout(4), interval = 100) {
-      app_wait_for_js(self, private, script = script, timeout = timeout, interval = interval)
+    wait_for_js = function(
+      script,
+      timeout = missing_arg(),
+      interval = 100
+    ) {
+      app_wait_for_js(
+        self, private,
+        script = script,
+        timeout = timeout,
+        interval = interval
+      )
     },
 
 
@@ -1497,51 +1511,6 @@ AppDriver <- R6Class( # nolint
     #' }
     get_variant = function() {
       app_get_variant(self, private)
-    },
-
-    #' @description
-    #' Get timeout value
-    #'
-    #' Retrieves the AppDriver's relative timeout value. The number of steps is multiplied by the internal `stepsize` value set in `$set_timeout(stepsize=)`.
-    #' @param steps A numeric multiplier to apply to the timeout value.
-    #' @return The number of milliseconds equal to `steps * stepsize` where `stepsize` is set from `$set_timeout(stepsize=)`
-    #' @examples
-    #' \dontrun{
-    #' app_path <- system.file("examples/01_hello", package = "shiny")
-    #'
-    #' app <- AppDriver$new(app_path)
-    #'
-    #' app$get_timeout()
-    #' #> [1] 1000
-    #' app$get_timeout(5)
-    #' #> [1] 5000
-    #' }
-    get_timeout = function(steps = 1) {
-      app_get_timeout(self, private, steps = steps)
-    },
-    #' @description
-    #' Set timeout value stepsize
-    #'
-    #' @param stepsize The number of milliseconds that one timeout _step_ should represent.
-    #' @examples
-    #' \dontrun{
-    #' app_path <- system.file("examples/01_hello", package = "shiny")
-    #'
-    #' app <- AppDriver$new(app_path, stepsize = 500)
-    #'
-    #' app$get_timeout()
-    #' #> [1] 500
-    #' app$get_timeout(5)
-    #' #> [1] 2500
-    #'
-    #' app$set_timeout(1000)
-    #' app$get_timeout()
-    #' #> [1] 1000
-    #' app$get_timeout(5)
-    #' #> [1] 5000
-    #' }
-    set_timeout = function(stepsize) {
-      app_set_timeout(self, private, stepsize = stepsize)
     },
 
     #' @description